Move mock data to fixtures/ folder
