+++ tests/auth.test.ts
@@ -4,7 +4,7 @@ import { Config } from '../src/utils/config'
 import authCommand from '../src/commands/auth'
 import { authEndpoint, graphcoolConfigFilePath, systemAPIEndpoint } from '../src/utils/constants'
 import TestAuthServer from '../src/api/TestAuthServer'
-import { testToken } from './mock_data/mockData'
+import { testToken } from './fixtures/mock_data'
 import { TestSystemEnvironment } from '../src/types'
 import TestOut from '../src/system/TestOut'
 import 'isomorphic-fetch'
