+++ tests/push.test.ts
@@ -8,7 +8,7 @@ import {systemAPIEndpoint, graphcoolProjectFileName, graphcoolConfigFilePath} fr
 import {
   mockedPushSchemaResponse, mockProjectFile2, mockProjectFile3,
   mockedPullProjectResponse1, mockedPushSchemaResponseError, mockModifiedProjectFile1, mockedPushSchema1ResponseError
-} from './mock_data/mockData'
+} from './fixtures/mock_data'
 import { TestSystemEnvironment } from '../src/types'
 import TestOut from '../src/system/TestOut'
